+++ src/crypto/mod.rs
@@ -6,7 +6,7 @@ mod response;
 
 pub use client::CryptoClient;
 pub use enums::{Loc, Sort, TimeFrame};
-pub use model::{Bar, Orderbook, Quote, Snapshot, Trade};
+pub use model::{Bar, Orderbook, OrderbookLevel, Quote, Snapshot, Trade};
 pub use request::{
     BarsRequest, LatestBarsRequest, LatestOrderbooksRequest, LatestQuotesRequest,
     LatestTradesRequest, QuotesRequest, SnapshotsRequest, TradesRequest,